Initial rough sketch
Idiom is a GUI that translates text from language A into language B. It
uses the Google Translate tools to do this.

Currently it can translate anything into anything so long as you want
Swedish to English. Instead of using the proper Google Translate API it
scrapes the JSON-like data that it sends to the Web page.

Depends on libcurl, JSON-Glib 1, and GTK+ 3.