Icons used by GTK+
This adds icons at different sizes in two forms (one with two letters
for larger sizes, and one with one letter for smaller sizes), plus the
source GIMP file.

The default icon for all windows uses the added icon. The About dialog
also uses it.

This requires the icons to be installed in a path that GTK+ searches as
a theme path. The easiest thing to do is:

    cd data && sudo make prefix=/usr/local install

You only need to do that each time an icon changes.